# docs/openapi/lxp-ollama-qwen-fileupload.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LXP Backoffice Policy Assistant (Document Upload + RAG)",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/documents/upload": {
      "post": {
        "summary": "Upload Document",
        "description": "PDF/DOCX 문서를 업로드하면 청킹 후 ChromaDB(+BM25)에 적재하여 RAG 검색 대상에 추가합니다.\n\n🚨 Excel(.xlsx/.xls), HWP(.hwp/.hwpx)는 아직 미지원입니다. (select_reason.md 8번 항목 참고)",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "get": {
        "summary": "List Documents",
        "description": "현재 RAG 저장소에 적재되어 있는 문서(출처)별 청크 개수를 조회합니다.",
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "type": "array",
                  "title": "Response List Documents Api V1 Documents Get"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Reset Documents",
        "description": "업로드된 문서를 모두 초기화합니다 (ChromaDB 컬렉션 + BM25 인메모리 캐시).",
        "operationId": "reset_documents_api_v1_documents_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/analyze-policy": {
      "post": {
        "summary": "Analyze Policy",
        "operationId": "analyze_policy_api_v1_analyze_policy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
      "DocumentInfo": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          }
        },
        "type": "object",
        "required": [
          "source",
          "chunk_count"
        ],
        "title": "DocumentInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PolicyRequest": {
        "properties": {
          "new_policy_text": {
            "type": "string",
            "title": "New Policy Text"
          }
        },
        "type": "object",
        "required": [
          "new_policy_text"
        ],
        "title": "PolicyRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}